Extrais un composant pour le tableau de sélection de services
...afin de le réutiliser dans la suppression de mesures associées
