+++ svelte/lib/listeMesures/modificationStatutPrecision/etapes/SecondeEtape.svelte
@@ -1,27 +1,13 @@
 <script lang="ts">
   import type { ReferentielStatut } from '../../../ui/types.d';
-  import { decode } from 'html-entities';
-  import TagStatutMesure from '../../../ui/TagStatutMesure.svelte';
-  import Tableau from '../../../ui/Tableau.svelte';
   import Infobulle from '../../../ui/Infobulle.svelte';
   import type { ServiceAssocie } from '../../mesureGenerale/modification/TiroirModificationMultipleMesuresGenerales.svelte';
+  import TableauSelectionServices from '../../kit/TableauSelectionServices.svelte';
 
   export let statuts: ReferentielStatut;
   export let modificationPrecisionUniquement: boolean;
   export let idsServicesSelectionnes: string[];
   export let servicesAssocies: ServiceAssocie[];
-
-  const optionsFiltrage = {
-    categories: [{ id: 'statut', libelle: 'Statuts' }],
-    items: [
-      { libelle: 'À définir', valeur: undefined, idCategorie: 'statut' },
-      ...Object.entries(statuts).map(([id, statut]) => ({
-        libelle: statut,
-        valeur: id,
-        idCategorie: 'statut',
-      })),
-    ],
-  };
 
   const doitEtreALaFin = (service: ServiceAssocie) =>
     !service.peutEtreModifie ||
@@ -36,84 +22,31 @@
     }
     return doitEtreALaFin(a) ? 1 : -1;
   });
+
+  const predicationDesactivation = (donnee: ServiceAssocie) =>
+    (modificationPrecisionUniquement && !donnee.mesure.statut) ||
+    !donnee.peutEtreModifie;
 </script>
 
 <span class="explication">
   Sélectionnez les services concernés par ces modifications. Les données
   existantes seront remplacées.
 </span>
-<Tableau
-  colonnes={[
-    { cle: 'nom', libelle: 'Nom du service' },
-    { cle: 'statut', libelle: 'Statut actuel' },
-    { cle: 'modalites', libelle: 'Précision actuelle' },
-  ]}
-  donnees={servicesOrdonnes}
-  configurationRecherche={{
-    champsRecherche: ['nomService', 'organisationResponsable'],
-  }}
-  configurationFiltrage={{ options: optionsFiltrage }}
-  configurationSelection={{
-    texteIndicatif: {
-      vide: 'Aucun service sélectionné',
-      unique: 'service sélectionné',
-      multiple: 'services sélectionnés',
-    },
-    champSelection: 'id',
-    predicatSelectionDesactive: (donnee) =>
-      (modificationPrecisionUniquement && !donnee.mesure.statut) ||
-      !donnee.peutEtreModifie,
-  }}
-  bind:selection={idsServicesSelectionnes}
+
+<TableauSelectionServices
+  {statuts}
+  bind:idsServicesSelectionnes
+  {servicesOrdonnes}
+  {predicationDesactivation}
 >
-  <svelte:fragment slot="cellule" let:donnee let:colonne>
-    {@const desactive =
-      !donnee.peutEtreModifie ||
-      (modificationPrecisionUniquement && !donnee.mesure.statut)}
-    {#if colonne.cle === 'nom'}
-      <div class="contenu-nom-service">
-        <div class="intitule-service" class:desactive>
-          <span class="nom">{decode(donnee.nomService)}</span>
-          <span class="organisation">{donnee.organisationResponsable}</span>
-        </div>
-        {#if !donnee.peutEtreModifie}
-          <Infobulle
-            contenu="Vous ne pouvez pas sélectionner ce service car vous ne disposez pas des droits d’écriture."
-          />
-        {/if}
-      </div>
-    {:else if colonne.cle === 'statut'}
-      <div class="conteneur-statut">
-        <div class:desactive>
-          <TagStatutMesure
-            referentielStatuts={statuts}
-            statut={donnee.mesure.statut}
-          />
-        </div>
-        {#if modificationPrecisionUniquement && !donnee.mesure.statut}
-          <Infobulle
-            contenu="Cette précision ne peut pas être appliquée à ce service, car il ne dispose pas actuellement d'un statut."
-          ></Infobulle>
-        {/if}
-      </div>
-    {:else if colonne.cle === 'modalites'}
-      {@const contenu = decode(donnee.mesure.modalites)}
-      {@const contenuTropLong = contenu.length > 90}
-      <div class="precision">
-        <span>{contenuTropLong ? contenu.slice(0, 90) + '...' : contenu}</span>
-        {#if contenuTropLong}
-          <img
-            src="/statique/assets/images/icone_voir_plus.svg"
-            alt={contenu}
-            title={contenu}
-            width="16px"
-            height="16px"
-          />
-        {/if}
-      </div>
+  <svelte:fragment slot="infoStatutMesure" let:donnee>
+    {#if modificationPrecisionUniquement && !donnee.mesure.statut}
+      <Infobulle
+        contenu="Cette précision ne peut pas être appliquée à ce service, car il ne dispose pas actuellement d'un statut."
+      ></Infobulle>
     {/if}
   </svelte:fragment>
-</Tableau>
+</TableauSelectionServices>
 
 <style lang="scss">
   span {
@@ -122,54 +55,7 @@
     line-height: 1.5rem;
   }
 
-  .contenu-nom-service {
-    display: flex;
-    flex-direction: row;
-    gap: 8px;
-    align-items: center;
-  }
-
-  .intitule-service {
-    display: flex;
-    flex-direction: column;
-    gap: 4px;
-    width: 318px;
-
-    .nom {
-      font-weight: bold;
-    }
-
-    .nom,
-    .organisation {
-      white-space: nowrap;
-      overflow: hidden;
-      text-overflow: ellipsis;
-    }
-  }
-
-  .conteneur-statut {
-    width: 148px;
-  }
-
-  .conteneur-statut {
-    display: flex;
-    flex-direction: row;
-    gap: 8px;
-    align-items: center;
-  }
-
-  .desactive {
-    opacity: 0.5;
-  }
-
   .explication {
     max-width: 500px;
-  }
-
-  .precision {
-    display: flex;
-    flex-direction: row;
-    gap: 8px;
-    align-items: center;
   }
 </style>
